Attempt to decompile ai00c7

+++ src/include/library/library.h
@@ -4,8 +4,7 @@
 
 u8 random(void);
 
-// Not sure what type the argument is
-u8 *ailistFindById(s32 ailistid);
+u8 *ailistFindById(u16 ailistid);
 
 s32 getCurrentStageId(void);
